+++ test/controlkeel/entry_point_test.exs
@@ -17,6 +17,8 @@ defmodule ControlKeel.EntryPointTest do
     previous_application_start_fun =
       Application.get_env(:controlkeel, :entry_point_application_start_fun)
 
+    previous_logger_handler = :logger.get_handler_config(:default)
+
     on_exit(fn ->
       if previous do
         System.put_env("__BURRITO", previous)
@@ -55,6 +57,14 @@ defmodule ControlKeel.EntryPointTest do
       else
         Application.delete_env(:controlkeel, :entry_point_application_start_fun)
       end
+
+      case previous_logger_handler do
+        {:ok, handler} ->
+          restore_default_logger_handler(handler)
+
+        _ ->
+          :ok
+      end
     end)
 
     :ok
@@ -73,6 +83,29 @@ defmodule ControlKeel.EntryPointTest do
     Application.put_env(:controlkeel, :cli_plain_arguments_provider, fn -> [~c"help"] end)
 
     assert CLI.standalone_argv() == ["help"]
+  end
+
+  test "standalone start redirects default logger handler to stderr" do
+    parent = self()
+
+    System.put_env("__BURRITO", "1")
+    Application.put_env(:controlkeel, :cli_plain_arguments_provider, fn -> [~c"help"] end)
+
+    configure_default_logger_type(:standard_io)
+
+    Application.put_env(:controlkeel, :entry_point_halt_fun, fn exit_code ->
+      send(parent, {:halted, exit_code})
+      :ok
+    end)
+
+    capture_io(fn ->
+      assert {:ok, _pid} = ControlKeel.EntryPoint.start(:normal, [])
+    end)
+
+    assert_receive {:halted, 0}
+
+    assert {:ok, %{config: config}} = :logger.get_handler_config(:default)
+    assert config[:type] == :standard_error
   end
 
   test "standalone help executes and halts synchronously" do
@@ -121,5 +154,31 @@ defmodule ControlKeel.EntryPointTest do
     assert_receive :application_started
     assert_receive {:executed, %{command: :init}}
     assert_receive {:halted, 0}
+  end
+
+  defp configure_default_logger_type(type) do
+    case :logger.get_handler_config(:default) do
+      {:ok, %{module: module} = handler} ->
+        replacement = %{handler | config: Map.put(handler.config || %{}, :type, type)}
+
+        with :ok <- :logger.remove_handler(:default),
+             {:ok, _handler_id} <- :logger.add_handler(:default, module, replacement) do
+          :ok
+        else
+          _ -> :ok
+        end
+
+      _ ->
+        :ok
+    end
+  end
+
+  defp restore_default_logger_handler(handler) do
+    with :ok <- :logger.remove_handler(:default),
+         {:ok, _handler_id} <- :logger.add_handler(:default, handler.module, handler) do
+      :ok
+    else
+      _ -> :ok
+    end
   end
 end